Started working on Neuronmancer 2.0
