+++ src/Makefile
@@ -307,3 +307,4 @@ clean:
 #	rm -rf ./bin/$(TARGET_ARCH)/$(TARGET_OS)/$(BUILD_TYPE)/neuronmancer
 
 clobber: clean
+
